+++ python/datamodel/validate/__init__.py
@@ -1,3 +1,4 @@
 
+from .add import *
 from .check import *
 from .models import *